Move combine_lookups into lookup.rs
